Sequence of dictionaries can be converted to a ROOT file automatically  (#100)
If a top level function is a Select whose result is a dictionary, it can be rentered automatically as a ROOT file
In this case, this removes the need for a AsROOTTTree function call.
The dictionary keys are used as column names.
Documentation and Tests were updated.
Added a new method called get_as_ROOT to do this work. Integrated it into the xaod code.